Refactor description panel
