+++ apps/frontend/src/components/questions/questionPage/QuestionWorkspace.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Split from "react-split";
-import DescriptionPanel from "./DescriptionPanel";
+import DescriptionPanel from "./descriptionPanel/DescriptionPanel";
 import { Question } from "@/database/question/entities/question.entity";
 import CodeEditor from "./codeEditor/CodeEditor";
 import ChatWidget from "@/components/chat/ChatWidget";
